+++ CHANGELOG.md
@@ -1121,6 +1121,7 @@ activity and on its child transactions (which can be actuals, refunds, and adjus
 - ISPF activities are not shown to users when the feature flag is enabled
 - ISPF activities are not shown in search results when the feature flag is enabled
 - Exclude ISPF-related entities from exports when the feature flag is enabled
+- The first step for an ISPF programme is now choosing between ODA and Non-ODA
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-120...HEAD
 [release-120]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-119...release-120